requirements.txt
Flask==2.3.3
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
joblib==1.3.2
python-dateutil==2.8.2
pytz==2023.3
scipy==1.11.2
six==1.16.0
threadpoolctl==3.2.0
requests==2.31.0
google-api-python-client==2.95.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.22.0 